Move public type definitions into roosterjs-content-model-type package
