+++ packages/roosterjs-content-model-api/lib/modelApi/block/toggleModelBlockQuote.ts
@@ -1,7 +1,6 @@
 import { areSameFormats, createFormatContainer, unwrapBlock } from 'roosterjs-content-model-dom';
 import { getOperationalBlocks, isBlockGroupOfType } from 'roosterjs-content-model-core';
 import { wrapBlockStep1, wrapBlockStep2 } from '../common/wrapBlock';
-import type { OperationalBlocks } from 'roosterjs-content-model-core';
 import type { WrapBlockStep1Result } from '../common/wrapBlock';
 import type {
     ContentModelBlock,
@@ -10,6 +9,7 @@ import type {
     ContentModelFormatContainer,
     ContentModelFormatContainerFormat,
     ContentModelListItem,
+    OperationalBlocks,
 } from 'roosterjs-content-model-types';
 
 /**
