+++ packages/roosterjs-content-model-core/lib/publicApi/selection/collectSelections.ts
@@ -1,7 +1,6 @@
 import { getClosestAncestorBlockGroupIndex } from '../model/getClosestAncestorBlockGroupIndex';
 import { isBlockGroupOfType } from '../model/isBlockGroupOfType';
 import { iterateSelections } from './iterateSelections';
-import type { IterateSelectionsOption } from './iterateSelections';
 import type {
     ContentModelBlock,
     ContentModelBlockGroup,
@@ -11,29 +10,11 @@ import type {
     ContentModelParagraph,
     ContentModelSegment,
     ContentModelTable,
+    IterateSelectionsOption,
+    OperationalBlocks,
     TableSelectionContext,
+    TypeOfBlockGroup,
 } from 'roosterjs-content-model-types';
-import type { TypeOfBlockGroup } from '../model/getClosestAncestorBlockGroupIndex';
-
-/**
- * Represent a pair of parent block group and child block
- */
-export type OperationalBlocks<T extends ContentModelBlockGroup> = {
-    /**
-     * The parent block group
-     */
-    parent: ContentModelBlockGroup;
-
-    /**
-     * The child block
-     */
-    block: ContentModelBlock | T;
-
-    /**
-     * Selection path of this block
-     */
-    path: ContentModelBlockGroup[];
-};
 
 /**
  * Get an array of selected parent paragraph and child segment pair
