+++ packages/roosterjs-content-model-core/lib/utils/paste/mergePasteContent.ts
@@ -5,7 +5,6 @@ import { domToContentModel } from 'roosterjs-content-model-dom';
 import { getSegmentTextFormat } from '../../publicApi/domUtils/getSegmentTextFormat';
 import { getSelectedSegments } from '../../publicApi/selection/collectSelections';
 import { mergeModel } from '../../publicApi/model/mergeModel';
-import type { MergeModelOption } from '../../publicApi/model/mergeModel';
 import type {
     BeforePasteEvent,
     ClipboardData,
@@ -13,6 +12,7 @@ import type {
     ContentModelDocument,
     ContentModelSegmentFormat,
     IEditor,
+    MergeModelOption,
 } from 'roosterjs-content-model-types';
 
 const EmptySegmentFormat: Required<ContentModelSegmentFormat> = {
