+++ packages/roosterjs-content-model-plugins/test/tableEdit/tableResizerTest.ts
@@ -1,4 +1,3 @@
-import * as TestHelper from '../TestHelper';
 import { getModelTable } from './tableData';
 import { TableEditPlugin } from '../../lib/tableEdit/TableEditPlugin';
 import {
